My_Portfolio/backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
whitenoise==6.6.0
python-dotenv==1.0.0
django-cors-headers==4.3.1